build: clear our 18 warnings, none of them by silencing
A clean build emitted 178 warnings, 18 of them ours; the rest are vendored and stay.

Two RAND_MAX sites were real bugs rather than style. 0x7fffffff is not representable
in float, so dividing by (float)RAND_MAX rounds the divisor up to 2^31 and the
quotient never quite reaches 1. gametest had seven of these and now shares one
rand01() that divides in double. veg_rand_range had the same defect with an explicit
cast in front of it -- the cast silenced the compiler's report of exactly this and
answered none of it.

postfx_sss_max_sigma_per_depth took a non-const mat4 and only ever read projection
[1][1]. That is our function, so the parameter is const now; render.c passes it from a
const Engine and needed no cast at all.

The remaining cglm sites cannot be fixed that way: glm_mat4_copy, glm_mat4_quat,
glm_mat4_ins3 and glm_vec3_muladds all read their source and declare it non-const, and
they are vendored. Casting there is admitting their signature is wrong; dropping const
from our own SceneNode and Bone pointers to match would have been the alternative and
is worse. draw_list.c and shadow.c already established the (vec4*) spelling.

Two postfx call sites were missing the (const float*) that every other mat4 uniform in
the file carries. Two dead functions in import.c are deleted -- extract_rotation_quat
had no references and compute_bone_global_rest only called itself.

Verified from scratch: 0 warnings in our sources, 24 goldens 0 px, forest arms
unchanged.

+++ cetra/src/procedural/vegetation_tex.c
@@ -464,8 +464,14 @@ static float leaf_half_width(float t, float skew) {
 
 // Texture-space randomness only. The tree generator keeps its own stream
 // (tree_gen.c) so shape never depends on how much texture work ran first.
+//
+// The division is in DOUBLE. RAND_MAX is 0x7fffffff, which float cannot represent, so
+// (float)RAND_MAX rounds the divisor UP to 2^31 and the quotient never quite reaches 1
+// -- the explicit cast that used to be here silenced the compiler's warning about
+// exactly that rather than answering it. double holds RAND_MAX exactly, so the only
+// rounding left is the single one on the way out.
 float veg_rand_range(float min_val, float max_val) {
-    return min_val + (float)rand() / (float)RAND_MAX * (max_val - min_val);
+    return min_val + (float)(rand() / (double)RAND_MAX) * (max_val - min_val);
 }
 
 // Secondary veins, as a 0..1 mask. Each one leaves the midrib at its own point
